[CORE] Move config.hpp to core/ and simplify error logs
