+++ engine/core/src/config.cpp
@@ -1,4 +1,4 @@
-#include "engine/core/config/config.hpp"
+#include "engine/core/config.hpp"
 
 #include "engine/core/base.hpp"
 #include "engine/core/config/config_loader.hpp"
